feat: rewrite in rust  and improve kftray server

+++ src-tauri/utils/kubeforward/src/kubecontext.rs
@@ -175,7 +175,7 @@ pub async fn list_service_ports(
                                             service_port_infos.push(KubeServicePortInfo {
                                                 name: cp.name.clone(),
                                                 port: Some(IntOrString::Int(
-                                                    cp.container_port as i32,
+                                                    cp.container_port,
                                                 )),
                                             });
                                             break 'port_search;
